Now has active initial frequency. need to figure out adding it to weight. Needs updating on selection in the future too.

+++ autoSuggest.h
@@ -3,5 +3,5 @@
 #include <vector>
 
 void closestWordSearch(std::vector<std::string> &combinations, std::vector<std::string> &closestResponses,
-    std::vector<int> &closestResponsesDist, std::vector<std::vector<std::string>>&dictGraph, const std::vector<std::vector<char>>&keyGraph, float &largestMinDist, float weight);
+    std::vector<float> &closestResponsesDist, std::vector<std::vector<std::string>>&dictGraph, const std::vector<std::vector<char>>&keyGraph, float &largestMinDist, float weight);
 std::vector<std::string> autoSuggest(sqlite3* db,std::string word,std::vector<std::vector<char>> &keyGraph);
