+++ resources/algorithms.h
@@ -3,6 +3,6 @@
 
 int indexOfFirstChar(std::string input);
 int indexOfChar(char input);
-int distBFS(char typedLetter, char dictionaryLetter, const std::vector<std::vector<char>>& keyGraph);
-void insertClosestMatch(std::vector<std::string>& closestResponses, std::vector<int>& closestResponsesDist,
-    const std::string& dictWord, int largestMinDist, float current); 
+float distBFS(char typedLetter, char dictionaryLetter, const std::vector<std::vector<char>>& keyGraph);
+void insertClosestMatch(std::vector<std::string>& closestResponses, std::vector<float>& closestResponsesDist,
+    const std::string& dictWord, float largestMinDist, float current);
